# Fix #1099 - Improve checksum performance