Pull request overview
---------------------

- Fix #1099 - Improve checksum performance

This is a minimal increase, mostly in the case where you are doing it from a std::string (and not a path or istream)

```
----------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations
----------------------------------------------------------------------
BM_CheckSumOld/8                   912 ns          912 ns       765421
BM_CheckSumOld/16                  979 ns          979 ns       714320
BM_CheckSumOld/64                 1169 ns         1169 ns       598045
BM_CheckSumOld/256                1956 ns         1956 ns       358035
BM_CheckSumOld/1024               5097 ns         5097 ns       137452
BM_CheckSumOld/4096              17747 ns        17746 ns        39564
BM_CheckSumOld/16384             68148 ns        68147 ns        10134
BM_CheckSumOld/65536            273542 ns       273540 ns         2599
BM_CheckSumOld/262144          1097493 ns      1097491 ns          639
BM_CheckSumOld/524288          2163546 ns      2163510 ns          322
BM_CheckSumOld_BigO               4.14 N          4.14 N    
BM_CheckSumOld_RMS                   1 %             1 %    
BM_CheckSumNew/8                   527 ns          526 ns      1337075
BM_CheckSumNew/16                  618 ns          618 ns      1175249
BM_CheckSumNew/64                  812 ns          812 ns       885931
BM_CheckSumNew/256                1645 ns         1645 ns       443998
BM_CheckSumNew/1024               4763 ns         4763 ns       146804
BM_CheckSumNew/4096              17456 ns        17455 ns        40330
BM_CheckSumNew/16384             68506 ns        68501 ns        10228
BM_CheckSumNew/65536            270522 ns       270506 ns         2552
BM_CheckSumNew/262144          1079197 ns      1079124 ns          649
BM_CheckSumNew/524288          2162980 ns      2162841 ns          320
BM_CheckSumNew_BigO               4.12 N          4.12 N    
BM_CheckSumNew_RMS                   0 %             0 %    
BM_CheckSumNewDirect/8             102 ns          102 ns      6852042
BM_CheckSumNewDirect/16            144 ns          144 ns      4835272
BM_CheckSumNewDirect/64            329 ns          329 ns      2133190
BM_CheckSumNewDirect/256          1055 ns         1055 ns       660191
BM_CheckSumNewDirect/1024         3959 ns         3958 ns       177071
BM_CheckSumNewDirect/4096        15527 ns        15527 ns        44995
BM_CheckSumNewDirect/16384       61844 ns        61841 ns        11264
BM_CheckSumNewDirect/65536      247638 ns       247615 ns         2829
BM_CheckSumNewDirect/262144     988530 ns       988469 ns          706
BM_CheckSumNewDirect/524288    1979162 ns      1979089 ns          354
BM_CheckSumNewDirect_BigO         3.77 N          3.77 N    
BM_CheckSumNewDirect_RMS             0 %             0 %    
```

![image](https://user-images.githubusercontent.com/5479063/114049361-ae796700-988b-11eb-9b3a-8461090f32d8.png)




### Pull Request Author

Add to this list or remove from it as applicable.  This is a simple templated set of guidelines.

 - [ ] Model API Changes / Additions
 - [ ] Any new or modified fields have been implemented in the EnergyPlus ForwardTranslator (and ReverseTranslator as appropriate)
 - [ ] Model API methods are tested (in `src/model/test`)
 - [ ] EnergyPlus ForwardTranslator Tests (in `src/energyplus/Test`)
 - [ ] If a new object or method, added a test in NREL/OpenStudio-resources: Add Link
 - [ ] If needed, added VersionTranslation rules for the objects (`src/osversion/VersionTranslator.cpp`)
 - [ ] Checked behavior in OpenStudioApplication, adjusted policies as needed (`src/openstudio_lib/library/OpenStudioPolicy.xml`)
 - [ ] Verified that C# bindings built fine on Windows, partial classes used as needed, etc.
 - [ ] All new and existing tests passes
 - [ ] If methods have been deprecated, update rest of code to use the new methods

**Labels:**

 - [ ] If change to an IDD file, add the label `IDDChange`
 - [ ] If breaking existing API, add the label `APIChange`
 - [ ] If deemed ready, add label `Pull Request - Ready for CI` so that CI builds your PR

### Review Checklist

This will not be exhaustively relevant to every PR.
 - [ ] Perform a Code Review on GitHub
 - [ ] Code Style, strip trailing whitespace, etc.
 - [ ] All related changes have been implemented: model changes, model tests, FT changes, FT tests, VersionTranslation, OS App
 - [ ] Labeling is ok
 - [ ] If defect, verify by running develop branch and reproducing defect, then running PR and reproducing fix
 - [ ] If feature, test running new feature, try creative ways to break it
 - [ ] CI status: all green or justified